Unify some shared utility code in functionality tests
